Build infrastructure and index page
This is a big commit and it does a lot of things but they
can be summarized in two parts:

1. Builds out the infrastructure
  - All the partials: nav, footer, socials
  - Buils out all the css components (cta-menu, btn, hero, etc.)
  - Adds base styles for HTML elements
  - CSS reset
  - Etc.

2. Builds out the index page - it now looks just like the original (slightly better)

Next Steps:
- The h and p tag element styling in cta-menu either needs to be moved
into base.css or we need to change these elements to <div> elements
with the custom styling in cta-menu.css and just reference the css
classes in the css declarations.

- None of the css utility classes in typography.css are being used,
instead font-family and font-weight styling is sprinkled everywhere.
Do we want to figure out a way to not have to sprinkle this styling
everywhere or do we want to just delete these utility classes?