feat: blog (initial implementation)
This creates the initial implementation for the blog. Includes a blog
index page and a show page as well. Still needs styling, some markup,
updated decap config, and possibly moving the blog show page layout
into a regular page instead?

Also need to look into how to add middleware so that when the blog
post markdown files are transformed into html, that we add styling
(css class) to img tags.