Style blog post draft previews
Unfortunately, we can't use the same HTML for blog posts within the
blog post preview pane. Therefore, this commit introduces a new
preview.css file that copies most of the styles that live in blog/blog.css.
These styles target the HTML elements that are in the preview pane.

We only want the preview pane to show the title and body of the blog post,
we don't want to show the excerpt because this is only displayed in the blog's
index page. We also don't want to display the date becuase it requires special
formatting that we're doing in the eleventy config and that's not available in
the preview pane - I didn't want to re-write this js formatting in two places.
Therefore, we must create a React component that serves as the template for
the preview pane. The custom css targets the elements within this template.

See:
https://decapcms.org/docs/customization/